Fixed several things: - Changed file and class names of window and scouttalker - First step of new AboutDialog class - Corrected copyright date - Added QApplication variables (name, organization, etc.)
