+++ src/ScoutTalker.cpp
@@ -1,5 +1,5 @@
-/* Scout Talker
- * Copyright (C) 2015 - 2017 by Diana Capela
+/* Scout Talker - Scouting codes, ciphers and encryption program
+ * Copyright (C) 2015 - 2018 by Diana Capela
  *                              João  Santos    (joaompssantos@gmail.com)
  *
  * This program is free software; you can redistribute it and/or modify
@@ -17,24 +17,22 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-#include "scouttalker.h"
-#include "window.h"
+#include "ScoutTalker.h"
+#include "MainWindow.h"
 #include <QChar>
 #include <QString>
 #include <QStringList>
 
-ScoutTalker::ScoutTalker(Window *w, QObject *parent) : QObject(parent)
-{
+ScoutTalker::ScoutTalker(MainWindow *w) {
     toEncodeString = new QString("");
     encodedString = new QString("");
-    window = w;
+    mainWindow = w;
 
-    connect(window, SIGNAL(startEncode()), this, SLOT(encodeSlot()));
+    connect(mainWindow, SIGNAL(startEncode()), this, SLOT(encodeSlot()));
 }
 
 // Method that clears the class strings
-void ScoutTalker::clearStrings()
-{
+void ScoutTalker::clearStrings() {
     toEncodeString->clear();
     encodedString->clear();
 }
@@ -44,24 +42,25 @@ QString ScoutTalker::removeAccents(QString string) {
     // Check if the diacritic letters string has already been created. If not the string is now populated.
     if (diacriticLetters_.isEmpty()) {
         diacriticLetters_ = QString::fromUtf8("ŠŒŽšœžŸ¥µÀÁÂÃÄÅÆÇÈÉÊËÌÍÎÏÐÑÒÓÔÕÖØÙÚÛÜÝßàáâãäåæçèéêëìíîïðñòóôõöøùúûüýÿ");
-        noDiacriticLetters_ << "S"<<"OE"<<"Z"<<"s"<<"oe"<<"z"<<"Y"<<"Y"<<"u"<<"A"<<"A"<<"A"<<"A"<<"A"<<"A"<<"AE"<<"C"<<"E"<<"E"<<"E"<<"E"<<"I"<<"I"<<"I"<<"I"<<"D"<<"N"<<"O"<<"O"<<"O"<<"O"<<"O"<<"O"<<"U"<<"U"<<"U"<<"U"<<"Y"<<"s"<<"a"<<"a"<<"a"<<"a"<<"a"<<"a"<<"ae"<<"c"<<"e"<<"e"<<"e"<<"e"<<"i"<<"i"<<"i"<<"i"<<"o"<<"n"<<"o"<<"o"<<"o"<<"o"<<"o"<<"o"<<"u"<<"u"<<"u"<<"u"<<"y"<<"y";
+        noDiacriticLetters_ << "S" << "OE" << "Z" << "s" << "oe" << "z" << "Y" << "Y" << "u" << "A" << "A" << "A" << "A"
+                            << "A" << "A" << "AE" << "C" << "E" << "E" << "E" << "E" << "I" << "I" << "I" << "I" << "D"
+                            << "N" << "O" << "O" << "O" << "O" << "O" << "O" << "U" << "U" << "U" << "U" << "Y" << "s"
+                            << "a" << "a" << "a" << "a" << "a" << "a" << "ae" << "c" << "e" << "e" << "e" << "e" << "i"
+                            << "i" << "i" << "i" << "o" << "n" << "o" << "o" << "o" << "o" << "o" << "o" << "u" << "u"
+                            << "u" << "u" << "y" << "y";
     }
 
     // Creates the output string
     QString output = "";
 
     // Loop that iterates through the input string and replaces diacritic letters with their non-diacritic equivalen
-    for (int i = 0; i < string.length(); i++)
-    {
+    for (int i = 0; i < string.length(); i++) {
         QChar ch = string[i];
         int dIndex = diacriticLetters_.indexOf(ch);
 
-        if (dIndex < 0)
-        {
+        if (dIndex < 0) {
             output.append(ch);
-        }
-        else
-        {
+        } else {
             QString replacement = noDiacriticLetters_[dIndex];
             output.append(replacement);
         }
@@ -71,71 +70,62 @@ QString ScoutTalker::removeAccents(QString string) {
 }
 
 // Method that manages the encoding process
-void ScoutTalker::encode()
-{
+void ScoutTalker::encode() {
     // Clears the class strings
     clearStrings();
 
     // Read the string to encode from the window object and remove diatritic letters
-    toEncodeString->append(removeAccents(window->getToEncodeString()));
+    toEncodeString->append(removeAccents(mainWindow->getToEncodeString()));
 
     // Gets current tab from the window object
-    Window::availableTabs currentTab = window->getTab();
+    MainWindow::availableTabs currentTab = mainWindow->getTab();
 
     switch (currentTab) {
-    case Window::ChineseCode:
-        // Do the procedures to perform the chinese code
-        encodeChineseCode();
-        break;
-    case Window::AngularCode:
-        // Do the procedures to perform the angular code
-        encodeAngularCode();
-        break;
-    case Window::ReverseAlphabet:
-        // Do the procedures to perform the reverse alphabet
-        encodeReverseAlphabet();
-        break;
-    default:
-        // Do the procedures to perform the chinese code
-        encodeChineseCode();
-        break;
+        case MainWindow::ChineseCode:
+            // Do the procedures to perform the chinese code
+            encodeChineseCode();
+            break;
+        case MainWindow::AngularCode:
+            // Do the procedures to perform the angular code
+            encodeAngularCode();
+            break;
+        case MainWindow::ReverseAlphabet:
+            // Do the procedures to perform the reverse alphabet
+            encodeReverseAlphabet();
+            break;
+        default:
+            // Do the procedures to perform the chinese code
+            encodeChineseCode();
+            break;
     }
 
     // Pushes the resulting string to the window object
-    window->setEncodedString(*encodedString);
+    mainWindow->setEncodedString(*encodedString);
 }
 
-void ScoutTalker::encodeChineseCode()
-{
+void ScoutTalker::encodeChineseCode() {
     encodedString->append(toEncodeString);
 }
 
-void ScoutTalker::encodeAngularCode()
-{
+void ScoutTalker::encodeAngularCode() {
     encodedString->append(toEncodeString);
 }
 
-void ScoutTalker::encodeReverseAlphabet()
-{
+void ScoutTalker::encodeReverseAlphabet() {
     int nChars = toEncodeString->count();
     encodedString->append(toEncodeString);
 
-    for (int i = 0; i < nChars; i++)
-    {
+    for (int i = 0; i < nChars; i++) {
         QChar currentCharacter = (*toEncodeString)[i];
 
-        if (currentCharacter.unicode() >= 0x41 && currentCharacter.unicode() <= 0x5A)
-        {
+        if (currentCharacter.unicode() >= 0x41 && currentCharacter.unicode() <= 0x5A) {
             (*encodedString)[i] = (QChar) (0x9B - currentCharacter.unicode());
-        }
-        else if (currentCharacter.unicode() >= 0x61 && currentCharacter.unicode() <= 0x7A)
-        {
+        } else if (currentCharacter.unicode() >= 0x61 && currentCharacter.unicode() <= 0x7A) {
             (*encodedString)[i] = (QChar) (0xDB - currentCharacter.unicode());
         }
     }
 }
 
-void ScoutTalker::encodeSlot()
-{
+void ScoutTalker::encodeSlot() {
     encode();
 }
